# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.APIStatus:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: boolean
    type: object
  model.Paging:
    properties:
      count:
        type: integer
      limit:
        type: integer
      offset:
        type: integer
    type: object
  model.Result:
    properties:
      apistatus:
        $ref: '#/definitions/model.APIStatus'
      data: {}
      paging:
        $ref: '#/definitions/model.Paging'
    type: object
  opensea.GetAssetsParams:
    description: Asset listing params
    properties:
      asset_contract_address:
        type: string
      asset_contract_addresses:
        items:
          type: string
        type: array
      collection:
        type: string
      collection_editor:
        type: string
      collection_slug:
        type: string
      cursor:
        type: integer
      include_orders:
        type: boolean
      limit:
        type: integer
      order_direction:
        type: string
      owner:
        type: string
      token_ids:
        items:
          type: integer
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: haluk.a.turan@gmail.com
    name: API Support
  description: |-
    fetches the assets by user wallet address
    one of the parameters walletAddress and x-wallet is mandatory and must be provided
  termsOfService: http://swagger.io/terms/
  title: NFT PORTAL API
  version: "1.0"
paths:
  /listAssets/{walletAddress}:
    post:
      consumes:
      - application/json
      parameters:
      - description: Wallet Address
        in: path
        name: walletAddress
        type: string
      - description: Wallet Address
        in: header
        name: x-wallet
        type: string
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: asset params
        in: body
        name: asset-params
        required: true
        schema:
          $ref: '#/definitions/opensea.GetAssetsParams'
      produces:
      - application/json
      responses:
        "200":
          description: status = true, Code = 200
          schema:
            $ref: '#/definitions/model.Result'
        "400":
          description: status = true, Code = 400, Message = Provide a valid walletAddress
          schema:
            $ref: '#/definitions/model.Result'
  /listAssetsTest/{walletAddress}:
    post:
      consumes:
      - application/json
      parameters:
      - description: Wallet Address
        in: path
        name: walletAddress
        type: string
      - description: Wallet Address
        in: header
        name: x-wallet
        type: string
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: asset params
        in: body
        name: asset-params
        required: true
        schema:
          $ref: '#/definitions/opensea.GetAssetsParams'
      produces:
      - application/json
      responses:
        "200":
          description: status = true, Code = 200
          schema:
            $ref: '#/definitions/model.Result'
        "400":
          description: status = true, Code = 400, Message = Provide a valid walletAddress
          schema:
            $ref: '#/definitions/model.Result'
swagger: "2.0"